chore(logging): add basic logging
