+++ lib/index.js
@@ -74,6 +74,7 @@ function applyLevel (context, spec, level) {
   }
 
   try {
+    console.log(`Applying level ${level} resources: ${names.join(', ')}`)
     var res = kubectl.apply(context, JSON.stringify(input))
   } catch (err) {
     err.message = `Failed to apply resources for level ${level}: ${err.message}`
@@ -95,7 +96,9 @@ async function deploy (options) {
     throw err
   }
 
+  console.log(`Applying namespaces: ${spec.namespaces.join(', ')}`)
   applyNamespaces(options.context, spec.namespaces)
+  console.log(`Applying pull secrets to ${Object.keys(spec.imagePullSecrets).join(', ')}`)
   applyImagePullSecrets(options.context, spec.imagePullSecrets)
   applyConfig(options.context, spec.configuration.concat(spec.secrets))
   for (const level of spec.levels) {
@@ -118,8 +121,11 @@ async function run (options) {
   const [name, namespace] = options.job.split('.')
   const job = spec.resources[options.job]
   const type = job.hasOwnProperty('cronJob') ? 'cronJob' : 'job'
+  console.log(`Deleting old ${type} "${name}"`)
   kubectl.del(options.context, namespace, type, name)
-  kubectl.apply(options.context, JSON.stringify(job[type]))
+  console.log(`Applying new ${type}`);
+  const result = kubectl.apply(options.context, JSON.stringify(job[type]))
+  console.log(`Created new ${type} "${result.metadata.name}"`)
 }
 
 module.exports = {
